nuke remaining warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@61342 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/sockets/php_sockets_win.c
@@ -52,7 +52,7 @@ ssize_t readv(SOCKET sock, const struct iovec *iov, int iovcnt) {
 	remain = bytes = (size_t) retval;
 	
 	for(i=0; i<iovcnt; i++) {
-		len = (iov[i].iov_len < remain) ? iov[i].iov_len : remain;
+		len = ((unsigned int)iov[i].iov_len < remain) ? iov[i].iov_len : remain;
 		memcpy(iov[i].iov_base, buffer+pos, len);
 		pos += len;
 		remain -= len;
